Bug fixed: Unrecognized bad lines
It seems that bad lines, which had a different number of columns than
others, were simply ignored in the last version. This bug was due to
pandas.read_csv and resolved by the update of pandas from 1.3 to 1.4.
Because of this bug fix of pandas, pyclx.pyclx.read_celex turned to be
able to detect bad lines and return an error. In order to accommodate
variable column numbers, depending on the number of rows to read, the
function pyclx.pyclx.read_celex was updated to the great extent. Now,
the function returns a pandas dataframe with the maximum number of
columns of the data being loaded. When some rows have fewer elements
(columns), missing values will be registered as empty strings.

On branch master
Changes to be committed:
      modified:   pyclx/__init__.py
      modified:   pyclx/pyclx.py
      new file:   tests/resources/gml.cd
      new file:   tests/resources/gml.readme
      modified:   tests/test_pyclx.py