Issue resolved: Data types
In the last version, all the columns were the object class. In the
current version, all the columns are converted to integers as long as it
is possible. For those that cannot be treated as integers, they are
untouched in the object class.

On branch master
Changes to be committed:
      modified:   pyclx/__init__.py
      modified:   pyclx/pyclx.py
      modified:   tests/test_pyclx.py